renamed components to multi-word
