+++ src/components/HeroBanner.vue
@@ -1,5 +1,4 @@
 <script setup>
-import { defineProps } from "vue";
 
 import getImageSize from '@/helpers/getImageSize';
